requirements.txt
astroid==1.6.1
bcrypt==3.1.4
beautifulsoup4==4.6.0
blinker==1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mongoengine==0.15.0
numpy==1.14.1
pandas==0.22.0
pycparser==2.18
PyJWT==1.5.3
pylint==1.8.2
pymongo==3.6.0
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
